% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_overlap.R
\name{block_overlap}
\alias{block_overlap}
\title{Compute the covariance overlap between each of the trajectory blocks from an input list
of eigensystems.}
\usage{
block_overlap(eigensys.list, eigfrom, eigto)
}
\arguments{
\item{eigensys.list}{A list of eigen-systems determined from comp_eigensystem()}

\item{eigfrom}{First eigenvector considered in computing the covariance overlap}

\item{eigto}{Last eigenvector considered in computing the covariance overlap}
}
\value{
A matrix containing the covariance overlap between each of the trajectory blocks
}
\description{
Compute the covariance overlap between each of the trajectory blocks from an input list
of eigensystems.
}
