% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_hubs.R
\name{identify_hubs}
\alias{identify_hubs}
\title{Identify the allosteric hubs from the ergodic sector mutual information matrix}
\usage{
identify_hubs(sector.list, n.hubs)
}
\arguments{
\item{sector.list}{A list of matrices, each is an averaged mutual information matrix for an ergodic sector}

\item{n.hubs}{An integer number of the number of top most significant hubs.}
}
\value{
A two-element list containing the top allosteric hubs and the associated mutual information couplings
between those allosteric hubs
}
\description{
This function computes the top n most signficant allosteric hub fragments. The function takes
the ergodic sector mutual information matrices from extract_sectors() and returns a list of the predicted top
most signficant allosteric hubs.
}
\details{
The returned hubs are 'structural alphabet fragments' rather than amino acid residues. This level of abstraction
results from encoding the MD trajectory into strings of structural fragments. The per-residue mutual information
can be obtained by parsing the results of this function to resid_MIreconstr().
}
