% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mats.R
\name{read_mats}
\alias{read_mats}
\title{Load the structural alphabet mutual information matrix trajectory.}
\usage{
read_mats(workingdir, matrix.suffix)
}
\arguments{
\item{workingdir}{Path to the input file (eg. "~/pnurse/MD_dir/")}

\item{matrix.suffix}{Suffix for mutual information matrix used as a pattern identifier (eg. "*nMImat.out")}
}
\value{
A list of mutual information matrices
}
\description{
This function loads a series of symmetric matrices into the R environment.
It assumes that all mutual information matrices are symmetric and are sorted
according to their temporal order.
}
