% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_str_traj.R
\name{read_traj_file}
\alias{read_traj_file}
\title{Read trajectory (XTC or DCD), return it as bio3d trajectory
Read structure (PDB or GRO), return it as bio3d structure}
\usage{
read_traj_file(workingdir, trajectory.name, traj.format = "xtc")
}
\arguments{
\item{workingdir}{Path to the input file}

\item{name}{of the input trajectory and structure template file}
}
\value{
bio3d structure trajectory
}
\description{
Read trajectory (XTC or DCD), return it as bio3d trajectory
Read structure (PDB or GRO), return it as bio3d structure
}
