% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_encode.R
\name{sa_encode}
\alias{sa_encode}
\title{sa_encode.R}
\usage{
sa_encode(traj.xyz, str.format = "pdb")
}
\arguments{
\item{pdbfile}{}
}
\value{
structural alphabet string
}
\description{
Encode a protein structure with the M32K25 structural alphabet.
}
\details{
Given a PDB structure this function will fit the C-alpha PDB coordinates
with the M32K25 structural alphabet, selecting the optimal fragment for each
successive four C-alpha residue set. The function returns the string of fitted
structural alphabet fragments as a string
}
