% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectors_3dplot.R
\name{sectors_3dplot}
\alias{sectors_3dplot}
\title{Image plot of the covariance overlap between all mutual information matrices defined in the input trajectory.}
\usage{
sectors_3dplot(overlap.matrix.s, t.traj)
}
\arguments{
\item{overlap.matrix.s}{Smoothed matrix of covariance overlaps, computed using matrix_smooth().}

\item{t.traj}{Length of the MD simulation, in nanoseconds.}
}
\value{
An image plot
}
\description{
This function generates and image plot of the smoothed covariance overlap matrix generated by matrix_smooth().
The function generates an image plot.
}
