% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_analyse.R
\name{testmat}
\alias{testmat}
\title{sa_analyse.R}
\usage{
testmat(num.row)
}
\arguments{
\item{sa.mat}{Alignment of structural alphabet strings}
}
\value{
structural alphabet string
}
\description{
[TT]g_sa_analyze[tt] performs advanced analyses on the alignment of structural,
 strings ([TT]-sa[tt]) produced by [TT]g_sa_encode[tt] (Pandini et al., FASEB J. [BB]26[bb], 868,
 (2012)).[PAR],
}
\details{
Option [TT]-MImatrix[tt] calculates the Mutual Information (MI) matrix between pairs,
 of columns (positions) in the alignment. The MI values are written to a ASCII,
 [TT].out[tt] file ([TT]-MImat[tt]). The normalized MI matrix ([TT]-nMImat[tt]), the joint entropy,
 matrix ([TT]-jHmat[tt]) and the expected finite size error matrix ([TT]-eeMImat[tt]) are also,
 calculated.[PAR],

The calculation of the statistically significance for the MI values can be requested,
 with [TT]-nSample[tt] > 0. A random background distribution of [TT]nSample[tt],
 samples is generated by shuffling the letters in each column of the,
 alignment. MI values averaged over the randomized samples ([TT]-meanMImat[tt]), the,
 associated standard deviations ([TT]-stdMImat[tt]), Z-scores ([TT]-ZMImat[tt]) and [TT]p-values[tt],
 ([TT]-pvalueMImat[tt]) are evaluated and written to ASCII [TT].out[tt] files.[PAR],

Option [TT]-trmat[tt] writes the transition probability matrix between pairs of SA,
 letters to an ASCII [TT].out[tt] file ([TT]-trans[tt]).[PAR],

A functional analysis can be requested by providing an [TT].xvg[tt] file with a,
 time-dependent function-related property ([TT]-fvalue[tt]). The MI values between,
 selected columns (positions) of the alignment and a discretized version of,
 the functional property are calculated. The positions in the alignment can be,
 selected providing a [TT].ndx[tt] file ([TT]-n[tt]). The column containing the functional,
 property in the [TT].xvg[tt] file can be specified with the [TT]-xvgcol[tt] flag. A summary,
 of MI statistics is written for each selected position to a separate [TT].log[tt],
 file ([TT]-MIlog[tt]). For visualization purposes, the discretized form of,
 the functional property and a numeric version of the structural string are,
 written for each selected position to a separate [TT].xvg[tt] file ([TT]-MIxvg[tt]) and to a,
 separate [TT].out[tt] file ([TT]-MIout[tt]).[PAR],

The same basic statistical analyses performed by [TT]g_sa_encode[tt] can also be,
 requested:[PAR],

Option [TT]-entropy[tt] writes the Shannon entropy of each position to an [TT].xvg[tt] file,
 ([TT]-Hlf[tt] for local fit or [TT]-Hgf[tt] for global fit).[PAR],

Option [TT]-profile[tt] writes the relative frequency of each letter at each position,
 to an ASCII [TT].dat[tt] file ([TT]-prolf[tt] for local fit or [TT]-progf[tt] for global fit).[PAR],

Option [TT]-xpm[tt] writes the alignment in [TT].xpm[tt] format: this can be converted to,
 postscript with [TT]xpm2ps[tt] for an immediate visualization of the time evolution,
 of the encoding for each position. The name of the [TT].xpm[tt] file can be specified,
 with the [TT]-xpm[tt] flag.
}
